ugc/tests/requirements.txt
aiohttp==3.9.5
backoff==2.2.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
gevent==24.2.1
gunicorn==22.0.0
kafka-python3==3.0.0
passlib==1.7.4
pydantic==2.6.3
pydantic-settings==2.2.1
pytest-flask==1.3.0
pytest-mock==3.14.0
requests==2.31.0
sentry-sdk[flask]==1.18.0
